releve enhancement (Ambulance revenu)
git-svn-id: http://emr.moh.gov.rw/svn/modules/moh-billing/branches/1.6.x@3186 6101c9c6-cf70-4d46-9f71-e9b75042c301

+++ web/src/org/openmrs/module/mohbilling/web/controller/MohBillingRevenueController.java
@@ -284,12 +284,12 @@ public class MohBillingRevenueController extends ParameterizableViewController {
 
 							if (category.equals("AUTRES")) {
 
-								double patientCost = item.getQuantity()
+								double patientCost = item.getQuantity().doubleValue()
 										* item.getUnitPrice().doubleValue();
 								patDueAmt = patDueAmt + patientCost;
 
 							} else {
-								double patientCost = item.getQuantity()
+								double patientCost = item.getQuantity().doubleValue()
 										* item.getUnitPrice().doubleValue()
 										* (100 - currentRate) / 100;
 								patDueAmt = patDueAmt + patientCost;
@@ -303,7 +303,7 @@ public class MohBillingRevenueController extends ParameterizableViewController {
 					if (svceCateg.startsWith("AMBULAN")) {
 						if (fspId == 96) {
 
-							double patientCost = item.getQuantity()
+							double patientCost = item.getQuantity().doubleValue()
 									* item.getUnitPrice().doubleValue()
 									* (100 - currentRate) / 100;
 
@@ -315,7 +315,7 @@ public class MohBillingRevenueController extends ParameterizableViewController {
 
 						if (docuLegIds.contains(fspId)) {
 
-							double patientCost = item.getQuantity()
+							double patientCost = item.getQuantity().doubleValue()
 									* item.getUnitPrice().doubleValue()
 									* (100 - currentRate) / 100;
 
@@ -330,7 +330,7 @@ public class MohBillingRevenueController extends ParameterizableViewController {
 
 						if (actMorgueIds.contains(fspId)) {
 
-							double patientCost = item.getQuantity()
+							double patientCost = item.getQuantity().doubleValue()
 									* item.getUnitPrice().doubleValue()
 									* (100 - currentRate) / 100;
